0913 중간커밋

+++ KingGgang_project/src/main/java/com/teamb/controller/MyController.java
@@ -77,35 +77,67 @@ public class MyController {
 	
 	@RequestMapping("/deletePro.my")
 	public String deletePro(HttpServletRequest req,HttpSession session) {
-		//로그인 세션이 없어서 테스트 불가 
-		//String id = String.valueOf(session.getAttribute("id"));
-		//String password = String.valueOf(session.getAttribute("password"));
-		String id =  req.getParameter("id");
-		String password = req.getParameter("password");
-		
+		String id = (String)session.getAttribute("mbId");
+		String passwd = (String)req.getAttribute("passwd");
+		System.out.println(id+passwd);
 		MemberDTO dto = myMapper.getMember(id);
-		String filename = dto.getProfile_name(); 
-		File file = new File(upLoadPath,filename);
+		int res = myMapper.checkPassword(id,passwd);
+		String msg=null,url=null;
+		switch(res){
+		case MemberDTO.OK:
+			int re = myMapper.deleteMember(dto.getMemberNum());
+				if(re>0){
+					msg="탈퇴 성공!!";
+					url="home.do";
+				}else{
+					msg="탈퇴 실패!! 관리자에게 문의하세요";
+					url="home.do";
+				}
+		case MemberDTO.NOT_ID:
+			msg="ID를 확인해주세요";
+			url="delete.my";
+		case MemberDTO.NOT_PW:
+			msg="비밀번호를 확인해주세요";
+			url="delete.my";
+		case MemberDTO.ERROR:
+			msg="에러!! 관리자에게 문의하세요";
+			url="delete.my";
+		}	
+		req.setAttribute("url", url);
+		req.setAttribute("msg", msg);
+		return "my/alert";
+	}
 		
-		int res = myMapper.deleteMember(id,password);
+		
+		/*String id =  req.getParameter("id");
+		String password = req.getParameter("password");
+		MemberDTO dto = myMapper.getMember(id);
+		int res = myMapper.deleteMember(id,password,dto.getMemberNum());
 		String msg = null, url=null;
-		if (res>0) {
-			if(file.delete()) {
-			url = "home.my";
-			msg = "delete successed";
-			}else {
-				url = "home.my";
-				msg = "delete successed but image is remained";
+		if (res > 0) {
+			if (dto.getProfile_name() == null) {
+				url = "memberList.mem";
+				msg = "delete successed";
+			} else {
+				String filename = dto.getProfile_name();
+				File file = new File(upLoadPath, filename);
+				if (file.delete()) {
+					url = "memberList.mem";
+					msg = "delete successed";
+				} else {
+					url = "memberList.mem";
+					msg = "delete successed but image is remained";
+				}
 			}
-		}else {
-			url = "home.my";
+		} else {
+			url = "memberList.mem";
 			msg = "delete failed";
 		}
 		
 		req.setAttribute("url", url);
 		req.setAttribute("msg", msg);
 		return "my/alert";
-	}
+	}*/
 	
 	@RequestMapping("/update.my")
 	public String updateForm(HttpServletRequest req) {
